backend/requirements.txt.txt
fastapi
uvicorn
langgraph
langchain-groq
python-dotenv
sqlalchemy
psycopg2-binary  # For PostgreSQL
pydantic